refactor(core): 抽取共享核心纯逻辑模块
